Update 20210704 Update Auto Date Update Script

+++ removeall.py
@@ -7,11 +7,9 @@ sys.path.append('/home/paullam/fyp/fyp/')
 os.environ['DJANGO_SETTINGS_MODULE'] = 'fyp.settings'
 django.setup()
 
-from datetime import datetime
-import pandas_datareader as dr
-import pandas as pd
 from candlesticks.models import Candlestick
 
 
 if __name__ == '__main__':
-    Candlestick.objects.all().delete()
+    if input('Are you sure to remove all candlesticks? (y/n)') == 'y':
+        Candlestick.objects.all().delete()
